Use requirements image with staticfiles built
